+++ packages/react/src/components/FileInput.tsx
@@ -1,32 +1,103 @@
-import React, { FunctionComponent, ChangeEvent, useState } from 'react'
+import React, {
+  FunctionComponent,
+  ChangeEvent,
+  useState,
+  SetStateAction,
+  Dispatch,
+  useRef,
+} from 'react'
 import styled from 'styled-components'
+import { icons } from './MainIcons'
+import { AadhaarQRValidation } from '../types'
+import { createBlobURL } from '../util'
 
 interface FileInputProps {
   onChange: (e: ChangeEvent<HTMLInputElement>) => void
   id: string
+  setQrStatus: Dispatch<SetStateAction<AadhaarQRValidation | null>>
+  qrStatus: AadhaarQRValidation | null
 }
 
 export const FileInput: FunctionComponent<FileInputProps> = ({
   onChange,
   id,
+  setQrStatus,
+  qrStatus,
 }) => {
-  const [fileName, setFileName] = useState<string>('No file selected')
+  const [fileName, setFileName] = useState<string | null>(null)
+  const fileInputRef = useRef<HTMLInputElement>(null)
+  const uploadIcon = createBlobURL(icons.fileUpload)
+  const xIcon = createBlobURL(icons.xBlack)
+
+  const clearFileInput = () => {
+    if (fileInputRef.current) {
+      fileInputRef.current.value = ''
+      setFileName(null)
+      setQrStatus(null)
+    }
+  }
+
   return (
-    <InputFile>
-      <UploadButton htmlFor={id}>Choose file</UploadButton>
-      <input
-        type="file"
-        id={id}
-        onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
-          if (!e.target.files) return
-          setFileName(e.target.files[0].name)
-          onChange(e)
-        }}
-        accept="image/*"
-        hidden
-      />
-      <FileName id="file-chosen">{fileName}</FileName>
-    </InputFile>
+    <>
+      <InputFile htmlFor={id}>
+        <input
+          type="file"
+          id={id}
+          ref={fileInputRef}
+          onChange={(e: React.ChangeEvent<HTMLInputElement>) => {
+            if (!e.target.files) return
+            setFileName(e.target.files[0].name)
+            onChange(e)
+          }}
+          accept="image/*"
+          hidden
+        />
+
+        <FileUploadIcon src={uploadIcon} />
+        <FileName id="file-chosen">Choose file</FileName>
+      </InputFile>
+      {fileName &&
+        (() => {
+          switch (qrStatus) {
+            case AadhaarQRValidation.ERROR_PARSING_QR:
+              return (
+                <>
+                  <InputFileWrong>
+                    <FileName id="file-chosen">{fileName}</FileName>
+                    <button onClick={clearFileInput}>
+                      <FileUploadIcon src={xIcon} />
+                    </button>
+                  </InputFileWrong>
+                  <DocumentResultWrong>Invalid QR Code.</DocumentResultWrong>
+                </>
+              )
+            case AadhaarQRValidation.SIGNATURE_VERIFIED:
+              return (
+                <>
+                  <InputFileCorrect>
+                    <FileName id="file-chosen">{fileName}</FileName>
+                    <button onClick={clearFileInput}>
+                      <FileUploadIcon src={xIcon} />
+                    </button>
+                  </InputFileCorrect>
+                  <DocumentResultCorrect>Valid QR Code.</DocumentResultCorrect>
+                </>
+              )
+            case AadhaarQRValidation.QR_CODE_SCANNED:
+              return (
+                <>
+                  <FileNameContainer>
+                    <FileName id="file-chosen">{fileName}</FileName>
+                    <button onClick={clearFileInput}>
+                      <FileUploadIcon src={xIcon} />
+                    </button>
+                  </FileNameContainer>
+                  <DocumentResult>Verifying QR Code.</DocumentResult>
+                </>
+              )
+          }
+        })()}
+    </>
   )
 }
 
@@ -34,30 +105,125 @@ const FileName = styled.span`
   margin-left: 5px;
 `
 
-const InputFile = styled.div`
+const InputFile = styled.label`
   display: flex;
   align-items: center;
   overflow: hidden;
   white-space: nowrap;
   text-overflow: ellipsis;
   max-width: 80%;
-  border-radius: 0.5rem;
+  border-radius: 4px;
   border-width: 1px;
-  border-color: #d1d5db;
+  border-color: black;
   max-width: 100%;
   font-size: '16px';
   line-height: 1.25rem;
   color: #111827;
-  background-color: #f9fafb;
+  padding-top: 6px;
+  padding-bottom: 6px;
+  padding-left: 14px;
+  padding-right: 14px;
   cursor: pointer;
   margin-top: 0.3rem;
 `
 
-const UploadButton = styled.label`
+const FileNameContainer = styled.label`
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  max-width: 80%;
+  border-radius: 4px;
+  border-width: 1px;
+  border-color: black;
+  max-width: 100%;
+  font-size: '16px';
+  line-height: 1.25rem;
   color: #111827;
-  background-color: #345c93;
-  color: white;
-  padding: 0.5rem;
-  font-family: sans-serif;
+  padding-top: 6px;
+  padding-bottom: 6px;
+  padding-left: 14px;
+  padding-right: 14px;
   cursor: pointer;
+  margin-top: 0.3rem;
+`
+
+const InputFileCorrect = styled.label`
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  max-width: 80%;
+  border-radius: 4px;
+  border-width: 2px;
+  border-color: #00bf06;
+  max-width: 100%;
+  font-size: '16px';
+  line-height: 1.25rem;
+  color: #111827;
+  padding-top: 6px;
+  padding-bottom: 6px;
+  padding-left: 14px;
+  padding-right: 14px;
+  cursor: pointer;
+  margin-top: 0.3rem;
+`
+
+const InputFileWrong = styled.label`
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+  width: 100%;
+  display: flex;
+  align-items: center;
+  overflow: hidden;
+  white-space: nowrap;
+  text-overflow: ellipsis;
+  max-width: 80%;
+  border-radius: 4px;
+  border-width: 2px;
+  border-color: #ef4444;
+  max-width: 100%;
+  font-size: '16px';
+  line-height: 1.25rem;
+  color: #111827;
+  padding-top: 6px;
+  padding-bottom: 6px;
+  padding-left: 14px;
+  padding-right: 14px;
+  cursor: pointer;
+  margin-top: 0.3rem;
+`
+
+export const FileUploadIcon = styled.img`
+  height: 1.5rem;
+`
+const DocumentResultCorrect = styled.div`
+  color: #00bf06;
+  position: absolute;
+  font-size: 0.875rem;
+  margin-top: 4px;
+`
+
+const DocumentResult = styled.div`
+  color: #717686;
+  position: absolute;
+  font-size: 0.875rem;
+  margin-top: 4px;
+`
+const DocumentResultWrong = styled.div`
+  color: #ef4444;
+  position: absolute;
+  font-size: 0.875rem;
+  margin-top: 4px;
 `